+++ packages/lib/src/standardActions/index.ts
@@ -1,2 +1,3 @@
 export * from "./arrayActions"
 export * from "./objectActions"
+export * from "./standaloneActions"