+++ packages/lib/src/standardActions/objectActions.ts
@@ -1,48 +1,46 @@
 import { remove, set } from "mobx"
+import { toTreeNode } from "../tweaker/tweak"
 import { assertIsObject } from "../utils"
-import { fnModel } from "./fnModel"
+import { standaloneAction } from "./standaloneActions"
 
-const _objectActions = fnModel<any>("mobx-keystone/objectActions").actions({
-  set(key: PropertyKey, value: any): void {
-    set(this, key, value)
-  },
-
-  assign(partialObject: any): void {
-    assertIsObject(partialObject, "partialObject")
-    const keys = Object.keys(partialObject)
-    for (const key of keys) {
-      set(this, key, (partialObject as any)[key])
-    }
-  },
-
-  delete(key: PropertyKey): boolean {
-    return remove(this, key)
-  },
-
-  call(methodName: PropertyKey, ...args: any[]): any {
-    return this[methodName](...args)
-  },
-})
+const namespace = "mobx-keystone/objectActions"
 
 export const objectActions = {
-  set: _objectActions.set as <T extends object, K extends keyof T>(
-    target: T,
-    key: K,
-    value: T[K]
-  ) => void,
+  set: standaloneAction(
+    `${namespace}::set`,
+    <T extends object, K extends keyof T>(target: T, key: K, value: T[K]): void => {
+      set(target, key, value)
+    }
+  ),
 
-  assign: _objectActions.assign as <T extends object>(target: T, partialObject: Partial<T>) => void,
+  assign: standaloneAction(
+    `${namespace}::assign`,
+    <T extends object>(target: T, partialObject: Partial<T>): void => {
+      assertIsObject(partialObject, "partialObject")
+      const keys = Object.keys(partialObject)
+      for (const key of keys) {
+        set(target, key, (partialObject as any)[key])
+      }
+    }
+  ),
 
-  delete: _objectActions.delete as <T extends object, K extends keyof T>(
-    target: T,
-    key: K
-  ) => boolean,
+  delete: standaloneAction(
+    `${namespace}::delete`,
+    <T extends object, K extends keyof T>(target: T, key: K): boolean => {
+      return remove(target, key as any)
+    }
+  ),
 
-  call: _objectActions.call as <T extends object, K extends keyof T>(
-    target: T,
-    methodName: K,
-    ...args: T[K] extends (...args: any[]) => any ? Parameters<T[K]> : never
-  ) => T[K] extends (...args: any[]) => any ? ReturnType<T[K]> : never,
+  call: standaloneAction(
+    `${namespace}::call`,
+    <T extends object, K extends keyof T>(
+      target: T,
+      methodName: K,
+      ...args: T[K] extends (...args: any[]) => any ? Parameters<T[K]> : never
+    ): T[K] extends (...args: any[]) => any ? ReturnType<T[K]> : never => {
+      return (target as any)[methodName](...args)
+    }
+  ),
 
-  create: _objectActions.create as <T>(data: T) => T,
+  create: <T extends object>(data: T): T => toTreeNode(data),
 }