+++ packages/lib/test/standardActions/arrayActions.test.ts
@@ -1,8 +1,9 @@
-import { arrayActions } from "../../src"
+import { arrayActions, isTreeNode } from "../../src"
 import "../commonSetup"
 
 test("typed array", () => {
   const arr = arrayActions.create([1, 2])
+  expect(isTreeNode(arr)).toBe(true)
   expect(arr[0]).toBe(1)
   expect(arr.length).toBe(2)
 
@@ -16,6 +17,7 @@ test("typed array", () => {
 
 test("untyped array", () => {
   const arr = arrayActions.create<any>([1, 2])
+  expect(isTreeNode(arr)).toBe(true)
   expect(arr[0]).toBe(1)
   expect(arr.length).toBe(2)